code: Implementado melhorias na UI da página de consultas, formularios e consertado problemas de rotas no backend
- Adicionado lógica de hover para exibir opções de edição e exclusão em ContentLine
- Integrado o componente Filters para seleção e filtragem por status de consultas
- Corrigido a aplicação de bordas no ContentLine para visibilidade aprimorada
- Refatorado o PrimaryButton para incluir funcionalidade de navegação para a rota de nova consulta
- Expandido o conjunto de dados de teste para incluir mais consultas e facilitar a verificação da UI

+++ src/backend/src/main/java/com/fisiomais/service/ConsultaService.java
@@ -56,7 +56,7 @@ public class ConsultaService {
         return consultaRepository.findByConfirmacao(status);
     }
 
-    public List<Consulta> getConsultasByPacienteId(Long pacienteId) {
+    public List<Consulta> getConsultasByPacienteId(Integer pacienteId) {
         return consultaRepository.findByPaciente_Id(pacienteId);
     }
